pyproject.toml
[project]
name = "pythonmockone"
version = "0.1"
description = "Learning Python Mocking"
authors = [{ name = "Evan Young", email = "cantis@gmail.com" }]

[tool.black]
skip-string-normalization = true
max-line-length = 120

[tool.pytest.ini_options]
addopts = [ "--import-mode=importlib" ]

[tool.poetry]
name = "pythonmockone"
version = "0.1.0"
description = "Learning Python Mocking"
authors = ["Evan Young <cantis@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alembic = "1.8.1"
black = "22.10.0"
psycopg2 = "^2.9.5"
pytest = "^7.2.0"
pytest-mock = "^3.10.0"
python-dotenv = "^0.21.0"
attrs = "^22.1.0"
click = "^8.1.3"
colorama = "^0.4.6"
greenlet = "^2.0.1"
iniconfig = "^1.1.1"
mako = "^1.2.3"
markupsafe = "^2.1.1"
packaging = "^21.3"
pathspec = "^0.10.1"
platformdirs = "^2.5.3"
pluggy = "^1.0.0"
pyparsing = "^3.0.9"
sqlalchemy = "^1.4.44"
tomli = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
